Sync to Betaflight/Betaflight

+++ src/main/drivers/rx/rx_a7105.c
@@ -54,7 +54,7 @@ void a7105extiHandler(extiCallbackRec_t* cb)
     }
 }
 
-void A7105Init (uint32_t id)
+void A7105Init(uint32_t id)
 {
     spiDeviceByInstance(RX_SPI_INSTANCE);
     rxIntIO = IOGetByTag(IO_TAG(RX_IRQ_PIN)); /* config receiver IRQ pin */
@@ -73,12 +73,12 @@ void A7105Init (uint32_t id)
     A7105WriteID(id);
 }
 
-void A7105Config (const uint8_t *regsTable, uint8_t size)
+void A7105Config(const uint8_t *regsTable, uint8_t size)
 {
     if (regsTable) {
-        uint32_t timeout = 1000;
+        unsigned timeout = 1000;
 
-        for (uint8_t i = 0; i < size; i++) {
+        for (unsigned i = 0; i < size; i++) {
             if (regsTable[i] != 0xFF) {
                 A7105WriteReg ((A7105Reg_t)i, regsTable[i]);
             }
@@ -98,7 +98,7 @@ void A7105Config (const uint8_t *regsTable, uint8_t size)
     }
 }
 
-bool A7105RxTxFinished (uint32_t *timeStamp) {
+bool A7105RxTxFinished(uint32_t *timeStamp) {
     bool result = false;
 
     if (occurEvent) {
@@ -112,22 +112,22 @@ bool A7105RxTxFinished (uint32_t *timeStamp) {
     return result;
 }
 
-void A7105SoftReset (void)
+void A7105SoftReset(void)
 {
     rxSpiWriteCommand((uint8_t)A7105_00_MODE, 0x00);
 }
 
-uint8_t A7105ReadReg (A7105Reg_t reg)
+uint8_t A7105ReadReg(A7105Reg_t reg)
 {
     return rxSpiReadCommand((uint8_t)reg | 0x40, 0xFF);
 }
 
-void A7105WriteReg (A7105Reg_t reg, uint8_t data)
+void A7105WriteReg(A7105Reg_t reg, uint8_t data)
 {
     rxSpiWriteCommand((uint8_t)reg, data);
 }
 
-void A7105Strobe (A7105State_t state)
+void A7105Strobe(A7105State_t state)
 {
     if (A7105_TX == state || A7105_RX == state) {
         EXTIEnable(rxIntIO, true);
@@ -156,7 +156,7 @@ void A7105WriteID(uint32_t id)
     rxSpiWriteCommandMulti((uint8_t)A7105_06_ID_DATA, &data[0], sizeof(data));
 }
 
-uint32_t A7105ReadID (void)
+uint32_t A7105ReadID(void)
 {
     uint32_t id;
     uint8_t data[4];
@@ -165,7 +165,7 @@ uint32_t A7105ReadID (void)
     return id;
 }
 
-void A7105ReadFIFO (uint8_t *data, uint8_t num)
+void A7105ReadFIFO(uint8_t *data, uint8_t num)
 {
     if (data) {
         if(num > 64) {
@@ -177,7 +177,7 @@ void A7105ReadFIFO (uint8_t *data, uint8_t num)
     }
 }
 
-void A7105WriteFIFO (uint8_t *data, uint8_t num)
+void A7105WriteFIFO(uint8_t *data, uint8_t num)
 {
     if (data) {
         if(num > 64) {
